Rearch the CLI Frontend moving from Proof-of-Concept

+++ pkg/kstrace/kstrace.go
@@ -3,6 +3,7 @@ package kstrace
 import (
 	"bytes"
 	"context"
+	"errors"
 	"os"
 
 	"fmt"
@@ -14,20 +15,23 @@ import (
 	log "github.com/sirupsen/logrus"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/cli-runtime/pkg/genericclioptions"
 
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 )
 
 type KStracer struct {
-	client         kubernetes.Interface
-	targetPod      *corev1.Pod
-	traceNamespace string
-	tracePod       *corev1.Pod
-	traceImage     string
-	containerPIDs  []int64
-	restConfig     *rest.Config
-	socketPath     string
+	client            kubernetes.Interface
+	targetPod         *corev1.Pod
+	traceNamespace    string
+	tracePod          *corev1.Pod
+	traceImage        string
+	containerPIDs     []int64
+	restConfig        *rest.Config
+	socketPath        string
+	collectionTimeout time.Duration
+	outputDirectory   string
 }
 
 type PrivilegedPodOptions struct {
@@ -40,18 +44,20 @@ type PrivilegedPodOptions struct {
 
 type Tracer interface {
 	Start() error
-	Stop() error
-	Cleanup() error
+	Cleanup()
 }
 
-func NewKStracer(clientset kubernetes.Interface, restConfig *rest.Config, traceImage string, targetPod *corev1.Pod, namespace string, socketPath string) *KStracer {
+func NewKStracer(clientset kubernetes.Interface, restConfig *rest.Config, traceImage string, targetPod *corev1.Pod, namespace string, socketPath string,
+	timeout time.Duration, outputDirectory string) Tracer {
 	straceObject := KStracer{
-		traceImage:     traceImage,
-		traceNamespace: namespace,
-		restConfig:     restConfig,
-		client:         clientset,
-		targetPod:      targetPod,
-		socketPath:     socketPath,
+		traceImage:        traceImage,
+		traceNamespace:    namespace,
+		restConfig:        restConfig,
+		client:            clientset,
+		targetPod:         targetPod,
+		socketPath:        socketPath,
+		collectionTimeout: timeout,
+		outputDirectory:   outputDirectory,
 	}
 
 	return &straceObject
@@ -129,8 +135,11 @@ func getPodDefinition(options PrivilegedPodOptions) *corev1.Pod {
 }
 
 func waitForPodRunning(clientset kubernetes.Interface, namespace string, pod string) error {
+	// TODO maybe this should be relative to the collection timeout?
+	timeout := time.Second * 30
+
 	checkPodState := func() bool {
-		// TODO WaitGroup
+
 		podStatus, err := clientset.CoreV1().Pods(namespace).Get(context.TODO(), pod, metav1.GetOptions{})
 		if err != nil {
 			return false
@@ -143,15 +152,61 @@ func waitForPodRunning(clientset kubernetes.Interface, namespace string, pod str
 		return false
 	}
 
-	// TODO Replace with WaitGroup
-	for {
+	for timeout > 0 {
+		log.Debugf("Waiting for Tracer Pod %q to become ready", pod)
 		if checkPodState() {
 			break
 		}
-		log.Debugf("Waiting for Tracer Pod %q to become ready", pod)
-		time.Sleep(time.Second)
+
+		// Sleep
+		time.Sleep(time.Second * 2)
+		timeout -= time.Second * 2
+	}
+
+	if timeout <= 0 {
+		return fmt.Errorf("tracer pod did not start correctly. review event objects from namespace %q relating to pod %q for more information", namespace, pod)
 	}
 	return nil
+}
+
+func (tracer *KStracer) getIOStream(container string) (*genericclioptions.IOStreams, error) {
+	var err error
+
+	// Special case for std-out
+	if tracer.outputDirectory == "-" {
+		return &genericclioptions.IOStreams{Out: os.Stdout, In: nil, ErrOut: os.Stderr}, nil
+	}
+
+	// Ensure trace folder is present
+	if _, err := os.Stat(tracer.outputDirectory); errors.Is(err, os.ErrNotExist) {
+		err = os.Mkdir(tracer.outputDirectory, 0775)
+		if err != nil {
+			log.Infof("Unable to create directory for the strace collection. %v", err)
+			return nil, err
+		}
+	}
+
+	// Ensure Pod-folder is present
+	podTraceFolder := fmt.Sprintf("%s%c%s", tracer.outputDirectory, os.PathSeparator, tracer.targetPod.Name)
+	if _, err := os.Stat(podTraceFolder); errors.Is(err, os.ErrNotExist) {
+		err = os.Mkdir(podTraceFolder, 0775)
+		if err != nil {
+			log.Infof("Unable to create directory for the strace collection. %v", err)
+			return nil, err
+		}
+	}
+	// Create file for container trace
+	fileWriter, err := os.Create(fmt.Sprintf("%s/%s/%s_strace.log", tracer.outputDirectory, tracer.targetPod.Name, container))
+	if err != nil {
+		log.Infof("Unable to create logfile for the strace collection. %v", err)
+		return nil, err
+	}
+
+	return &genericclioptions.IOStreams{
+		Out:    fileWriter,
+		ErrOut: fileWriter,
+		In:     nil,
+	}, nil
 }
 
 func (tracer *KStracer) Start() error {
@@ -179,9 +234,16 @@ func (tracer *KStracer) Start() error {
 	}
 
 	// Run Strace for collected containerPIDs
-	for _, containerPID := range tracer.containerPIDs {
+	for index, containerPID := range tracer.containerPIDs {
 		log.Debugf("Running strace on container %d", containerPID)
-		err = tracer.StartStrace(containerPID)
+
+		// Write to a file with the container name
+		iostream, err := tracer.getIOStream(tracer.targetPod.Spec.Containers[index].Name)
+		if err != nil {
+			return err
+		}
+
+		err = tracer.StartStrace(containerPID, iostream)
 		if err != nil {
 			return err
 		}
@@ -189,11 +251,6 @@ func (tracer *KStracer) Start() error {
 
 	log.Info("Strace complete")
 	return err
-}
-
-func (tracer *KStracer) Stop() error {
-	return nil
-
 }
 
 func (tracer *KStracer) CreateStracePod(ctx context.Context, options PrivilegedPodOptions) (*corev1.Pod, error) {
@@ -217,24 +274,27 @@ func (tracer *KStracer) CreateStracePod(ctx context.Context, options PrivilegedP
 	return createdPod, nil
 }
 
-func (tracer *KStracer) StartStrace(targetPID int64) error {
-	// TODO Replace with other stream
-	stdOut := os.Stdout
-	stdErr := os.Stderr
-
+func (tracer *KStracer) StartStrace(targetPID int64, iostreams *genericclioptions.IOStreams) error {
 	command := fmt.Sprintf("strace -fp %d", targetPID)
+
+	// Configure Command Timeout
+	if tracer.collectionTimeout != 0 {
+		command = fmt.Sprintf("timeout -s 2 --preserve-status %f %s", tracer.collectionTimeout.Seconds(), command)
+	}
+
 	log.Infof("Running command %q inside pod %q", command, tracer.tracePod.Name)
 
 	execRequest := ExecRequest{
 		Client: tracer.client, RestConfig: tracer.restConfig, PodName: tracer.tracePod.Name,
-		Namespace: tracer.tracePod.Namespace, Command: command, Stdin: nil,
-		Stdout: stdOut, Stderr: stdErr, TTY: false,
+		Namespace: tracer.tracePod.Namespace, Command: command, TTY: false, IOStreams: iostreams,
 	}
 	exitCode, err := ExecCommand(execRequest)
 
-	if exitCode != 0 {
+	// NOTE: ExitCode 130 is the response from strace after getting `Ctrl+C`
+	if exitCode != 0 && exitCode != 130 {
 		return fmt.Errorf("the function has failed with exit code: %d", exitCode)
 	}
+	log.Infof("Strace command for Pod %q complete", tracer.tracePod.Name)
 
 	if err != nil {
 		return err
@@ -242,22 +302,25 @@ func (tracer *KStracer) StartStrace(targetPID int64) error {
 
 	return nil
 }
-func (tracer *KStracer) Cleanup() error {
+func (tracer *KStracer) Cleanup() {
 	ctx := context.TODO()
 	// Delete Pod
+	if tracer == nil || tracer.tracePod == nil {
+		return
+	}
+
 	err := tracer.client.CoreV1().Pods(tracer.tracePod.Namespace).Delete(ctx, tracer.tracePod.Name, metav1.DeleteOptions{})
 	if err != nil {
 		log.Fatalf("unable to delete strace pod %q from namespace %q. manual deletion is required.", tracer.tracePod.Name, tracer.tracePod.Namespace)
-		return err
 	}
 	tracer.tracePod = nil
-	return nil
 }
 
 func (tracer *KStracer) FindPodPIDs() ([]int64, error) {
 	// Specific to Crictl
-	stdOut := new(bytes.Buffer)
-	stdErr := new(bytes.Buffer)
+	iostreams := &genericclioptions.IOStreams{
+		In: nil, Out: new(bytes.Buffer), ErrOut: new(bytes.Buffer),
+	}
 
 	// Get all Container IDs for Pod
 	containerPIDs := []int64{}
@@ -271,8 +334,7 @@ func (tracer *KStracer) FindPodPIDs() ([]int64, error) {
 
 		execRequest := ExecRequest{
 			Client: tracer.client, RestConfig: tracer.restConfig, PodName: tracer.tracePod.Name,
-			Namespace: tracer.tracePod.Namespace, Command: command, Stdin: nil,
-			Stdout: stdOut, Stderr: stdErr, TTY: false,
+			Namespace: tracer.tracePod.Namespace, Command: command, IOStreams: iostreams, TTY: false,
 		}
 		exitCode, err := ExecCommand(execRequest)
 		if exitCode != 0 || err != nil {
@@ -280,13 +342,15 @@ func (tracer *KStracer) FindPodPIDs() ([]int64, error) {
 		}
 
 		// Process JSON
-		json := stdOut.Bytes()
+		// TODO perform checks against casting
+		json := iostreams.Out.(*bytes.Buffer).Bytes()
+
 		containerPID, err := jsonparser.GetInt(json, "info", "pid")
 		if err != nil {
 			return nil, err
 		}
 
-		log.Infof("Container PID %q found for Pod %q", containerPID, containerID)
+		log.Infof("Container PID %d found for Container %q", containerPID, containerID)
 		containerPIDs = append(containerPIDs, containerPID)
 
 	}
